requirements.txt
torch==2.4.0
torchvision==0.20.1
Pillow==9.4.0
numpy==1.24.0



